Moved the Group part in the controller to the GroupController.java. Modified methods in the GroupManager.java so that it uses GUID instead of group. (Add/Remove groups are still using group because I may need to modify codes in the UserManager, or maybe the UserController which is not yet implemented.) Add a few javadocs so that every methods in the GroupController.java and GroupManager.java have descriptions.
